requirements.txt
# Core App + UI
streamlit==1.46.1
streamlit-autorefresh

# OpenAI + Retry Safety
openai>=1.0.0
tenacity==9.1.2

# Document Handling
python-docx==1.2.0
docx-mailmerge
jinja2==3.1.6
lxml==6.0.0

# Excel / Data Handling
pandas==2.3.0
openpyxl==3.1.5
python-dateutil

# API Requests
requests==2.32.4
dropbox

# Config and Secrets
python-dotenv

# OCR / Document Conversion (optional)
PyMuPDF
pytesseract
pypandoc
pandoc


